Init commit on original PQL implementation
